+++ packages/koot/React/styles.js
@@ -1,6 +1,13 @@
-const styleMap = {}
+import React from 'react'
 
-export const checkAndWriteIntoHead = () => {
+export const StyleMapContext = React.createContext({})
+
+/**
+ * 将样式表写入到 head 标签内
+ * @param {Object} styleMap 
+ */
+export const checkAndWriteIntoHead = (styleMap = {}) => {
+    if (typeof styleMap !== 'object') return
     Object.keys(styleMap).forEach(wrapper => {
         const style = styleMap[wrapper]
         if (style.count > 0) {
@@ -20,11 +27,17 @@ export const checkAndWriteIntoHead = () => {
     })
 }
 
-export const append = (style) => {
+/**
+ * 追加样式
+ * @param {Object} styleMap 
+ * @param {Object|Array} style 
+ */
+export const append = (styleMap = {}, style) => {
     if (Array.isArray(style))
-        return style.forEach(theStyle => append(theStyle))
+        return style.forEach(theStyle => append(styleMap, theStyle))
 
-    // console.log(style)
+    if (typeof style !== 'object') return
+
     if (!styleMap[style.wrapper]) {
         styleMap[style.wrapper] = {
             css: style.css,
@@ -34,17 +47,58 @@ export const append = (style) => {
         styleMap[style.wrapper].count++
     }
 
-    if (__CLIENT__)
-        checkAndWriteIntoHead()
+    if (__CLIENT__) {
+        checkAndWriteIntoHead(styleMap)
+    }
 }
 
-export const remove = (style) => {
+/**
+ * 移除样式
+ * @param {Object} styleMap 
+ * @param {*} style 
+ */
+export const remove = (styleMap = {}, style) => {
     if (Array.isArray(style))
         return style.forEach(theStyle => remove(theStyle))
+
+    if (typeof style !== 'object') return
 
     if (styleMap[style.wrapper]) {
         styleMap[style.wrapper].count--
     }
 }
 
-export const get = () => styleMap
+const idDivStylesContainer = '__KOOT_ISOMORPHIC_STYLES_CONTAINER__'
+
+/**
+ * 分析 HTML 代码，解析已有样式表，将其从 HTML 代码中移除，并返回可以直接写入到 head 标签内的样式表代码
+ * @param {String} html 
+ * @returns {String} htmlStyles
+ */
+export const parseHtmlForStyles = (html) => {
+    const matches = html.match(new RegExp(`<div id="${idDivStylesContainer}">(.+)</div>`, 'm'))
+    return {
+        html: html.substr(0, matches.index),
+        htmlStyles: matches[1]
+    }
+}
+
+/**
+ * React 组件: 样式表内容容器
+ */
+export class StylesContainer extends React.Component {
+    static contextType = StyleMapContext
+    render() {
+        return (
+            <div
+                id={idDivStylesContainer}
+                dangerouslySetInnerHTML={{
+                    __html: Object.keys(this.context)
+                        .filter(id => !!this.context[id].css)
+                        .map(id => `<style id="${id}">${this.context[id].css}</style>`)
+                        .join('')
+                }}
+            />
+        )
+    }
+}