+++ packages/koot/core/webpack/config/create/module-rules.js
@@ -68,7 +68,14 @@ module.exports = (options = {}) => {
             (env === 'dev' && stage === 'client')
         ) ? true : false
 
-        const useSpCssLoader = `sp-css-loader?length=4&mode=replace&readable=${env === 'dev' ? 'true' : 'false'}`
+        const useSpCssLoader = {
+            loader: require.resolve('../../loaders/css'),
+            options: {
+                length: 4,
+                mode: 'replace',
+                readable: env === 'dev' ? true : false
+            }
+        }
         const useUniversalAliasLoader = {
             loader: "universal-alias-loader",
             options: {