+++ packages/koot/core/webpack/loaders/css/stats.js
@@ -1,0 +1,4 @@
+module.exports = {
+    sameIndex: 0,
+    collection: []
+}